requirements.txt
google-cloud-aiplatform
google-cloud-speech
flask